[FIX] - Added phpdocs on every controllers, services & requests, renamed controllers, services & requests and dispatched function in three services rather  than one
