+++ app/Services/System.php
@@ -1,0 +1,147 @@
+<?php
+
+namespace App\Services;
+
+use Illuminate\Support\Facades\Process;
+use Illuminate\Process\ProcessResult;
+
+/**
+ * Class System
+ * 
+ * Service that manages system operations
+ * 
+ * @package App\Services
+ */
+class System
+{
+    protected string $pythonPath;
+    protected string $scriptsPath;
+
+    public function __construct()
+    {
+        $this->pythonPath = env('PYTHON_PATH', '/usr/bin/python3');
+        $this->scriptsPath = base_path('scripts');
+    }
+
+    /**
+     * Execute a command as a specific user.
+     *
+     * @param  string  $command   The command to execute
+     * @return ProcessResult
+     * 
+     * @throws \RuntimeException If the user session is not found or the command fails
+     */
+    public function execute(string $command): ProcessResult
+    {
+        $user = session('vps_user');
+
+        if (!$user) {
+            throw new \RuntimeException('No user session found.');
+        }
+
+        $script = escapeshellarg($this->scriptsPath . '/execute.py');
+        $userArg = escapeshellarg($user);
+        $commandArg = escapeshellarg($command);
+
+        return Process::run("{$this->pythonPath} {$script} {$userArg} {$commandArg}");
+    }
+
+    /**
+     * Retrieve information about a folder.
+     *
+     * @param  string  $path   Path to the folder
+     * @return array{inode:int, size:int, updated_at:\DateTimeImmutable, created_at:\DateTimeImmutable}
+     *
+     * @throws \RuntimeException If the system command fails or returns an unexpected format
+     */
+    public function getFolderInfo(string $path): array
+    {
+        $escapedPath = escapeshellarg($path);
+
+        // On récupère inode, taille, date de modif Unix timestamp et date de création Unix timestamp
+        $cmd = sprintf('stat -c "%%i %%s %%Y %%W" %s', $escapedPath);
+        $result = $this->execute($cmd);
+
+        if (! $result->successful()) {
+            throw new \RuntimeException(
+                "Impossible d'exécuter stat sur « {$path} » : " . $result->errorOutput()
+            );
+        }
+
+        $output = trim($result->output());
+        $parts  = preg_split('/\s+/', $output, 4);
+
+        if (count($parts) !== 4 || ! ctype_digit($parts[0]) || ! ctype_digit($parts[1])) {
+            throw new \RuntimeException("Format de sortie inattendu : « {$output} »");
+        }
+
+        [$inode, $size, $updatedTs, $createdTs] = $parts;
+
+        return [
+            'path'       => $path,
+            'inode'      => (int) $inode,
+            'size'       => (int) $size,
+            'updated_at' => (new \DateTimeImmutable())->setTimestamp((int) $updatedTs),
+            'created_at' => (new \DateTimeImmutable())->setTimestamp((int) $createdTs),
+        ];
+    }
+
+    /**
+     * Retrieve a list of folders in the /projects directory.
+     * 
+     * @return array<string> List of folder paths
+     */
+    public function getFolders(): array
+    {
+        $result = $this->execute('ls /projects');
+
+        $output = trim($result->output());
+
+        if ($result->successful() && !empty($output)) {
+            return array_values(array_filter(
+                array_map(fn($folder) => "/projects/{$folder}", explode("\n", $output))
+            ));
+        }
+
+        return [];
+    }
+
+    /**
+     * Create a folder.
+     *
+     * @param  string  $path   The folder path to create
+     * @return ProcessResult   The result of the folder creation process
+     */
+    public function createFolder(string $path): ProcessResult
+    {
+        return $this->execute("sudo mkdir -p " . escapeshellarg($path));
+    }
+
+    /**
+     * Delete a folder.
+     *
+     * @param  string  $path        The folder path to delete
+     * 
+     * @throws \RuntimeException    If the folder does not exist or cannot be deleted
+     * 
+     * @return ProcessResult        The result of the folder deletion process
+     */
+    public function deleteFolder(string $path): ProcessResult
+    {
+        $availability = $this->execute("test -d " . escapeshellarg($path));
+
+        if (!$availability->successful()) {
+            throw new \RuntimeException("Folder does not exist.");
+        }
+
+        if ($path === '/') {
+            throw new \RuntimeException("Cannot delete root directory.");
+        }
+
+        if (dirname($path) === '/') {
+            throw new \RuntimeException("Cannot delete top-level system directory.");
+        }
+
+        return $this->execute("sudo rm -rf " . escapeshellarg($path));
+    }
+}
